[DO NOT MERGE] drivers: imx: EDMA driver
There are two ways to use the ESAI -- either manually put the samples in
ETDR or read them from ERDR, or use the EDMA0 controller to have it do
this for you. This commit implements the EDMA driver so I can do the
latter and use the ESAI in an easy way.

Signed-off-by: Paul Olaru <paul.olaru@nxp.com>